requirements_lock.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    bazel run //:deps.update
#
--extra-index-url https://pypi.python.org/simple/

absl-py==2.0.0 \
    --hash=sha256:9a28abb62774ae4e8edbe2dd4c49ffcd45a6a848952a5eccc6a49f3f0fc1e2f3 \
    --hash=sha256:d9690211c5fcfefcdd1a45470ac2b5c5acd45241c3af71eed96bc5441746c0d5
    # via -r requirements.in
bazel-runfiles==0.27.0 \
    --hash=sha256:ff2081001e27085a4c57bb828202603d3f3c349d3e9683c44aedf02836e8deab
    # via -r requirements.in
